feat: use macro for GetLastError logging
